Fix "vanishing beans" problem
Previously, we didn't update the bounding box and sphere
of the `InstancedMesh`s.
Consequently, when we moved the camera such that it could not see the
original bounding box, the beans would disappear
(despite the fact that they had since moved to another location,
and thus invalidated the original bounding box).
Now, we update the bounding box and sphere each frame.