Deploy from nearest tower to cursor
Because of the imprecise control over the camera angle,
it is hard to deploy from a tower that is far away
(since small mouse/trackpad movements result in large
groun cursor movements).
So rather than require the cursor be within the bounds of
a tower, we simply select the nearest friendly tower to the
cursor.